ACS-3723: Moving endpoints to private API path.
